Cleanup ChainedBFT
